Address PR review feedback: CSS polish and move UI files to contrib/ghostCell
- Fix header alignment when description expands (flex-start instead of center)
- Remove hover transition on ghost cell container (instant feedback)
- Add hover state to highlighted segmented toggle button
- Remove border on Dismiss button hover (opacity only)
- Use positronActionBar-selectBorder for Accept button separator
- Increase model indicator font size to 12px
- Move GhostCell UI files into contrib/ghostCell alongside controller
