+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookComponent.tsx
@@ -18,7 +18,7 @@ import { NotebookCodeCell } from './notebookCells/NotebookCodeCell.js';
 import { NotebookMarkdownCell } from './notebookCells/NotebookMarkdownCell.js';
 import { NotebookRawCell } from './notebookCells/NotebookRawCell.js';
 import { DeletionSentinel } from './notebookCells/DeletionSentinel.js';
-import { GhostCell } from './notebookCells/GhostCell.js';
+import { GhostCell } from './contrib/ghostCell/GhostCell.js';
 import { IEditorOptions } from '../../../../editor/common/config/editorOptions.js';
 import { FontMeasurements } from '../../../../editor/browser/config/fontMeasurements.js';
 import { PixelRatio } from '../../../../base/browser/pixelRatio.js';
