+++ src/vs/workbench/contrib/positronNotebook/browser/contrib/ghostCell/GhostCellInfoModalDialog.css
